requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
scikit-learn==1.4.2
pandas==2.2.2
joblib==1.4.2
numpy==1.26.4
flask-cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
flask-talisman==1.1.0
flask-compress==1.14
flask-limiter==3.5.0
psutil==5.9.8
